refactor and add readme
