+++ scripts/data_uploader/upload_ichikara.py
@@ -1,19 +1,26 @@
 import json
-from typing import Optional, Any
+import argparse
+from typing import Optional
 import wandb
 import weave
 from weave import Dataset
 
-# Initialize Weave
-weave.init('wandb-japan/ichikara-test')
+def parse_arguments():
+    """Parse command-line arguments."""
+    parser = argparse.ArgumentParser(description="Weave dataset uploader with wandb integration.")
+    parser.add_argument("--entity", type=str, required=True, help="WandB entity name")
+    parser.add_argument("--project", type=str, required=True, help="WandB project name")
+    parser.add_argument("--question_file", type=str, required=True, help="Path to the JSON file containing the questions")
+    
+    return parser.parse_args()
 
-def load_questions(question_file: str, begin: Optional[int], end: Optional[int]):
+def load_questions(question_file: str):
     """Load questions from a file."""
     with open(question_file, "r") as ques_file:
         try:
             data = json.load(ques_file)
             if isinstance(data, list):
-                questions = data[begin:end] if begin is not None and end is not None else data
+                questions = data
             else:
                 questions = [data]
         except json.JSONDecodeError as e:
@@ -22,11 +29,12 @@ def load_questions(question_file: str, begin: Optional[int], end: Optional[int])
 
     return questions
 
-# Load questions
-questions = load_questions("/workspace/filtered_ichikara_instruction_eval.json", None, None)
-
-# Create a dataset
-dataset = Dataset(name='ichikara_100', rows=questions)
-
-# Publish the dataset
-weave.publish(dataset)
+def main():
+    args = parse_arguments()
+    weave.init(f"{args.entity}/{args.project}")
+    questions = load_questions(args.question_file)
+    dataset = Dataset(name='ichikara', rows=questions)
+    weave.publish(dataset)
+    
+if __name__ == "__main__":
+    main()
