--- scripts/evaluator/ichikara_dataset_prep.py
@@ -1,119 +1,0 @@
-import json
-
-# Load the original data
-with open('/workspace/ichikara-instruction-eval-001-001.json', 'r', encoding='utf-8') as infile:
-    data = json.load(infile)
-
-# Specify the list of IDs to filter
-specified_ids = [
-    "ichikara-instruction-eval-001-001-0000997-001",
-    "ichikara-instruction-eval-001-001-0000980-001",
-    "ichikara-instruction-eval-001-001-0000979-001",
-    "ichikara-instruction-eval-001-001-0000969-001",
-    "ichikara-instruction-eval-001-001-0000958-001",
-    "ichikara-instruction-eval-001-001-0000954-001",
-    "ichikara-instruction-eval-001-001-0000937-001",
-    "ichikara-instruction-eval-001-001-0000934-001",
-    "ichikara-instruction-eval-001-001-0000884-001",
-    "ichikara-instruction-eval-001-001-0000877-001",
-    "ichikara-instruction-eval-001-001-0000869-001",
-    "ichikara-instruction-eval-001-001-0000867-001",
-    "ichikara-instruction-eval-001-001-0000822-001",
-    "ichikara-instruction-eval-001-001-0000814-001",
-    "ichikara-instruction-eval-001-001-0000812-001",
-    "ichikara-instruction-eval-001-001-0000807-001",
-    "ichikara-instruction-eval-001-001-0000802-001",
-    "ichikara-instruction-eval-001-001-0000800-001",
-    "ichikara-instruction-eval-001-001-0000780-001",
-    "ichikara-instruction-eval-001-001-0000776-001",
-    "ichikara-instruction-eval-001-001-0000773-001",
-    "ichikara-instruction-eval-001-001-0000763-001",
-    "ichikara-instruction-eval-001-001-0000753-001",
-    "ichikara-instruction-eval-001-001-0000745-001",
-    "ichikara-instruction-eval-001-001-0000744-001",
-    "ichikara-instruction-eval-001-001-0000703-001",
-    "ichikara-instruction-eval-001-001-0000659-001",
-    "ichikara-instruction-eval-001-001-0000628-001",
-    "ichikara-instruction-eval-001-001-0000626-001",
-    "ichikara-instruction-eval-001-001-0000607-001",
-    "ichikara-instruction-eval-001-001-0000595-001",
-    "ichikara-instruction-eval-001-001-0000575-001",
-    "ichikara-instruction-eval-001-001-0000574-001",
-    "ichikara-instruction-eval-001-001-0000571-001",
-    "ichikara-instruction-eval-001-001-0000570-001",
-    "ichikara-instruction-eval-001-001-0000569-001",
-    "ichikara-instruction-eval-001-001-0000565-001",
-    "ichikara-instruction-eval-001-001-0000563-001",
-    "ichikara-instruction-eval-001-001-0000559-001",
-    "ichikara-instruction-eval-001-001-0000557-001",
-    "ichikara-instruction-eval-001-001-0000556-001",
-    "ichikara-instruction-eval-001-001-0000547-001",
-    "ichikara-instruction-eval-001-001-0000546-001",
-    "ichikara-instruction-eval-001-001-0000545-001",
-    "ichikara-instruction-eval-001-001-0000544-001",
-    "ichikara-instruction-eval-001-001-0000541-001",
-    "ichikara-instruction-eval-001-001-0000535-001",
-    "ichikara-instruction-eval-001-001-0000533-001",
-    "ichikara-instruction-eval-001-001-0000526-001",
-    "ichikara-instruction-eval-001-001-0000501-001",
-    "ichikara-instruction-eval-001-001-0000496-001",
-    "ichikara-instruction-eval-001-001-0000491-001",
-    "ichikara-instruction-eval-001-001-0000485-001",
-    "ichikara-instruction-eval-001-001-0000476-001",
-    "ichikara-instruction-eval-001-001-0000473-001",
-    "ichikara-instruction-eval-001-001-0000470-001",
-    "ichikara-instruction-eval-001-001-0000469-001",
-    "ichikara-instruction-eval-001-001-0000461-001",
-    "ichikara-instruction-eval-001-001-0000459-001",
-    "ichikara-instruction-eval-001-001-0000458-001",
-    "ichikara-instruction-eval-001-001-0000455-001",
-    "ichikara-instruction-eval-001-001-0000453-001",
-    "ichikara-instruction-eval-001-001-0000452-001",
-    "ichikara-instruction-eval-001-001-0000448-001",
-    "ichikara-instruction-eval-001-001-0000446-001",
-    "ichikara-instruction-eval-001-001-0000444-001",
-    "ichikara-instruction-eval-001-001-0000441-001",
-    "ichikara-instruction-eval-001-001-0000434-001",
-    "ichikara-instruction-eval-001-001-0000431-001",
-    "ichikara-instruction-eval-001-001-0000429-001",
-    "ichikara-instruction-eval-001-001-0000428-001",
-    "ichikara-instruction-eval-001-001-0000423-001",
-    "ichikara-instruction-eval-001-001-0000421-001",
-    "ichikara-instruction-eval-001-001-0000418-001",
-    "ichikara-instruction-eval-001-001-0000403-001",
-    "ichikara-instruction-eval-001-001-0000398-001",
-    "ichikara-instruction-eval-001-001-0000383-001",
-    "ichikara-instruction-eval-001-001-0000379-001",
-    "ichikara-instruction-eval-001-001-0000371-001",
-    "ichikara-instruction-eval-001-001-0000369-001",
-    "ichikara-instruction-eval-001-001-0000355-001",
-    "ichikara-instruction-eval-001-001-0000305-001",
-    "ichikara-instruction-eval-001-001-0000289-001",
-    "ichikara-instruction-eval-001-001-0000268-001",
-    "ichikara-instruction-eval-001-001-0000187-001",
-    "ichikara-instruction-eval-001-001-0000173-001",
-    "ichikara-instruction-eval-001-001-0000166-001",
-    "ichikara-instruction-eval-001-001-0000146-001",
-    "ichikara-instruction-eval-001-001-0000122-001",
-    "ichikara-instruction-eval-001-001-0000121-001",
-    "ichikara-instruction-eval-001-001-0000119-001",
-    "ichikara-instruction-eval-001-001-0000113-001",
-    "ichikara-instruction-eval-001-001-0000112-001",
-    "ichikara-instruction-eval-001-001-0000109-001",
-    "ichikara-instruction-eval-001-001-0000059-001",
-    "ichikara-instruction-eval-001-001-0000049-001",
-    "ichikara-instruction-eval-001-001-0000046-001",
-    "ichikara-instruction-eval-001-001-0000045-001",
-    "ichikara-instruction-eval-001-001-0000030-001",
-    "ichikara-instruction-eval-001-001-0000023-001"
-]
-
-
-# Filter data by the specified IDs
-filtered_data = [entry for entry in data if entry["ID"] in specified_ids]
-
-# Save the filtered data to a new file
-with open('/workspace/ichikara100.json', 'w', encoding='utf-8') as outfile:
-    json.dump(filtered_data, outfile, ensure_ascii=False, indent=2)
-
-print(f"Filtered data has been saved to filtered_ichikara_instruction_eval.json")
